requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
fastapi==0.135.1
h11==0.16.0
idna==3.11
pillow==12.1.1
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
python-multipart==0.0.22
redis==7.3.0
starlette==0.52.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.41.0
